moving fetching NFTs by block expolorer to the same controller
