+++ src/main/kotlin/id/walt/nftkit/rest/PolkadotUniqueNftController.kt
@@ -85,4 +85,41 @@ object PolkadotUniqueNftController {
         it.description("Fetched NFT metadata")
     }
 
+    fun fetchparachainNFTS(ctx: Context) {
+        val chain = ctx.pathParam("chain")
+        val account = ctx.pathParam("account")
+
+
+        val result = PolkadotNftService.fetchAccountTokensBySubscan(PolkadotParachain.valueOf(chain), account)
+        ctx.json(result)
+    }
+
+    fun fetchparachainNFTSDocs() = document().operation {
+        it.summary("Fetching Tokens by Subscan")
+            .operationId("fetchparachainNFTS")
+            .addTagsItem(TAG)
+    }.pathParam<String>("chain") {
+        it.schema<PolkadotParachain> {}
+    }.pathParam<String>("account") {
+        it.schema<String> {}
+
+    }
+
+    fun fetchEvmErc721CollectiblesBySubscan(ctx: Context) {
+        val chain = ctx.pathParam("chain")
+        val account = ctx.pathParam("account")
+        val result = PolkadotNftService.fetchEvmErc721CollectiblesBySubscan(PolkadotParachain.valueOf(chain), account)
+        ctx.json(result)
+    }
+
+    fun fetchEvmErc721CollectiblesBySubscanDocs() = document().operation {
+        it.summary("Fetching EvmErc721 Collectibles by Subscan")
+            .operationId("fetchEvmErc721CollectiblesBySubscan")
+            .addTagsItem(TAG)
+    }.pathParam<String>("chain") {
+        it.schema<PolkadotParachain> {}
+    }.pathParam<String>("account") {
+        it.schema<String> {}
+
+    }
 }
